_StringBuffer: Ensure capacity isn’t less than requested in the ASCII case
In the ASCII case, the newly allocated buffer’s capacity was set to
the actual allocated capacity - 2, which may be 1 less than what was requested.

Fix this by removing capacityBump. The actual capacity will be the requested capacity rounded up to the nearest even number (plus any allocation overhead).

It looks like capacityBump was intended to guarantee that there’s always an extra hidden byte at the end of any ASCII buffer; however, that seems to be unnecessary now. (This is unrelated to nul termination.)