StringGuts: Annotate ARC hacks with effects(readonly).
When hacking around ARC by using opaque function calls on unsafe
bitpatterns, self still has to be guaranteed in case of arbitrary
effects. This can limit the degree to which the optimizer is allowed
to optimize. Add effects(readonly). As a result, this speeds up (and
even completely trivializes) many benchmarks.