[32bit] Use a custom enum type instead of AnyObject? in _StringObject
This allows us to eliminate the _StringGuts._extraBits stored property, reducing the size of String to three words on 32-bit platforms.

In exchange, we don't have uniform flag bits on 32-bit platforms any more, likely reducing performance of our existing _StringObject accessors. (The flag bits are now encoded in the enum representation in the tagged cases.)