[String] Scalar-alignment bug fixes.
Fixes a general category (pun intended) of scalar-alignment bugs
surroudning Substrings, and one specific one for the UTF16View.

Substrings formed from non-scalar-aligned indices exhibit many
behaviors as though they were scalar aligned according to SE-0180, but
this can be difficult to guarantee. Particularly vexing is that
Substring would like to rely on the validly-encoded invariant of
native Strings, but this cannot happen if formed with a sub-scalar
UTF-8 index. So, we scalar-align indices on Substring construction and
establish alignment as an invariant of Substring, eliminating
significant bug potential.

The runtime overhead of this is unfortunate, but is the necessary
result of SE-0180. String.Index is already paying a perf cost as well.

The UTF16View must scalar-align any indices from the UTF8View, and
this adds an alignment check to _getNativeOffset.